+++ src/main/java/api/client/UserClient.java
@@ -1,0 +1,43 @@
+package api.client;
+
+import api.model.UserCredentials;
+import io.qameta.allure.Step;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.http.ContentType;
+import io.restassured.response.ValidatableResponse;
+import io.restassured.specification.RequestSpecification;
+
+import static io.restassured.RestAssured.given;
+
+public class UserClient {
+    private static final String BASE_URL = "https://stellarburgers.nomoreparties.site/";
+    private static final String PATH_LOGIN = "/api/auth/login";
+    private static final String PATH_USER_UPDATE_DELETE = "/api/auth/user";
+
+    public RequestSpecification getSpec() {
+        return new RequestSpecBuilder()
+                .setContentType(ContentType.JSON)
+                .setBaseUri(BASE_URL)
+                .build();
+    }
+    @Step("Авторизация пользователя")
+    public ValidatableResponse loginUser(UserCredentials userCredentials) {
+        return given()
+                .spec(getSpec())
+                .body(userCredentials)
+                .when()
+                .post(PATH_LOGIN)
+                .then();
+    }
+
+    @Step("Удаление пользователя")
+    public ValidatableResponse deleteUser(String accessToken) {
+        return given()
+                .spec(getSpec())
+                .body(accessToken)
+                .when()
+                .delete(PATH_USER_UPDATE_DELETE)
+                .then();
+
+    }
+}
