+++ src/main/java/api/model/UserCredentials.java
@@ -1,0 +1,27 @@
+package api.model;
+
+public class UserCredentials {
+    private String email;
+    private String password;
+
+    public UserCredentials(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
